[cherry-pick] Cherry-pick 2.56.1 mobilenet 64028 64031 (#64591)
The mobilenet is not available in the torch hub anymore.
Of the two examples in question, one was untested (not included in BUILD
file).
We should remove both as they are 6 years old legacy v1 stack code at
this point.
(cherry picked from commit
https://github.com/ray-project/ray/commit/8d87a60eabac5fa33c2352be0d5dab3e0ab8ce62)


Upgrades the mobilenet example to mobilenet v3 and uses torchvision to
get the model to get around failures like this one:

https://buildkite.com/ray-project/postmerge/builds/17990#019eb515-1f34-4e8e-b177-5ec2cd7a6d50

`cannot import name '_cast_Long' from 'torch.onnx.symbolic_opset9'

(/opt/miniforge/lib/python3.12/site-packages/torch/onnx/symbolic_opset9.py)`
(cherry picked from commit
https://github.com/ray-project/ray/commit/d2f5d18700649823ecd94dfed14af6cdab6845c1)

---------
